requirements.txt
streamlit
Pillow
numpy
PyWavelets